Update investigating-lost-objects.adoc
telnet to Port 1402 (LDR console) has to be done towards the storage node that contains errors/lost objects. The step before you log in to root@admin-node and as there no new ssh login is mentioned, the telnet connection has to be made to the storage node to access the LDR service.